Add a new API to retrieve some commitIds
The last, the 10th, the 100th, the 200th and the 1000th.
This API is useful in order to prevent a full new sync in the case of
the last commitId is no longer known by the server. It happens for
example when the server has garbage collected some entries. If the
client is connecting after a lot of days, weeks, it's necessary to check
for older commitId.